jwtkms: Add a KMS partial mock and jwtkms tests
This commit adds a new `internal/mockkms` package which provides an
implementation of the KMSClient interface backed by in-memory storage.
The mock isn't as strict as the real KMS, but does implement the Sign,
Verify, and GetPublicKey methods faithfully enough to allow for testing
of jwtkms.

This commit also introduces tests for all of the jwtkms signing methods,
using the mockkms package to provide the backing implementation.